requirements.txt
requests
beautifulsoup4
lxml
tqdm
python-dotenv
